Split into new code file and header file
